refactor(frontend/scripting): introduce ScriptExecutionEnvironment and use AsyncDisposable pattern
